Move UnionFind into a separate file to reduce the clutter
